+++ src/sveltekit/templates.ts
@@ -31,6 +31,9 @@ import * as Sentry from '@sentry/sveltekit';
 Sentry.init({
   dsn: '${dsn}',
   tracesSampleRate: 1.0,
+
+  // uncomment the line below to enable Spotlight (https://spotlightjs.com)
+  // spotlight: import.meta.env.DEV,  
 });
 
 // If you have custom handlers, make sure to place them after \`sentryHandle()\` in the \`sequence\` function.