+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- feat(nextjs): Pin installed Next.js SDK version to version 7 (#550)
 - Add note about `tunnelRoute` and Next.js middleware incompatibility (#544)
 - feat(nextjs): Added comment to add spotlight in Sentry.init for Next.js server config (#545)
 
@@ -25,8 +26,10 @@
 
 ## 3.20.1
 
-- fix(nextjs): Replace deprecated Sentry API calls in example page templates (#520)
-- fix(sveltekit): Replace deprecated Sentry API calls in example page templates (#520)
+- fix(nextjs): Replace deprecated Sentry API calls in example page templates
+  (#520)
+- fix(sveltekit): Replace deprecated Sentry API calls in example page templates
+  (#520)
 
 ## 3.20.0